Split Tasty testers into different modules
